Add inter-procedural argument update analysis.
Coupled with on-going alias analysis, this make it possible to add a
const qualifier to any argument *not* updated by a function. Not
critical in terms of performance, although it prunes a few copies.
However the const are now added whenever it seems legal and the whole
validation suite is ok with this, which gives pretty good guarantee over
the correctness of the analysis.