Defensive approach in the pass manager when apply() on unknown pass type
I still don't know if we should raise or warn or whatever here. At least
do not call blindly display() unless there is a super-class for passes
that provide it.
Defensive programmign apart, code is more readable with the check for
'Analysis' type. Still wonder if "transformation" is the right name
for the pass.