Auto conversion from tuple to list when possible
The basic idea is to *always* rely on core::make_tuple to create tuple
and on decltype(core::make_tuple(...)) to refer to the type of a tuple.

The one can overload core::make_tuple to return a ltuple instead of a
tuple when all parameters have the same type. A ltuple is just a list
with another type. So it is iterable etc and behaves fine with pythonic.
And the extra type information makes it possible to convert it to a
tuple when going back to the Python world .

This commit closes #70.