Add more info to Pythran spec error
Now it displays the line number where the error happens, and the symbol that causes it. The filename is also set in the SyntaxError, but i'm not sure it has any effect.

I tried adding the column and the content of the line, to see if there would be a clever print, but it just displayed the same message, so that code wasn't added to the commit.

For reference: http://www.dabeaz.com/ply/ply.html contains the code I used to find the line number of the error, and the code to find the column of the token