Add math functions for python long ints
Currently in test_long.py, two tests fail: test_atan2 and test_hypot. My guess is that it has to do with the different argument types. Those tests have been commented out.

For math functions with multiple arguments, I resorted to a single template call (instead of testing each argument to see if it's a mpz_class). As such, it creates a conflict with a few math functions who already have such template calls, and so their 'using' statement have been commented out.

For those functions, it might be wiser to keep the using instead and do 2 template calls (one with the first argument as mpz_class, the second with the second argument as mpz_class) and one non-template call (both arguments as mpz_class).