Correct handling of the -e switch
A regression led to the generation of BOOST_PYTHON code with this flag,
which is incorrect. Improve dic/CLI.rst to document the interface and
add a few extra test cases to prevent such regressions.

Add pythran-config script similar to python-config to help third party
compilation. Pythran-config is preferred over pkg-config because it can
directly call the pythran package, and it is aware of pythranrc.