Add basic SIMD testing
This is *not* to have full coverage, but rather to prevent regression
from what is already supported, as no testing is done otherwise. As a
consequence, some tests are commented in test_simd.py.

Note that this forces the inclusion of a large part of Boost.SIMD, and
thus the tests are running slower, (but only if boost.simd is activated
and numpy is used)