Simplify the nested_container mechanics
The one based on is_iterable trait was too hard to for both G++ and
clang++ (I actually had a version supported by g++ and a version
supported by clang++), this one is simple, but requires some manual
bookkeeping.

The main issue was that is_iterable is called to compute the type depth,
including the one of list<T>, and list is not completely defined while
we ask for its iterablity.