Implement a caching mechanism for the passmanager
*really* faster compilation thanks to this (almost twice as fast for
nqueens) as it avoids redundant computations when a transformation
depends on two analyse that both depend on the same analyse.

This required fine-tuning the self.update fields for various
transformations.

Also, constant folding used to replace nodes with equivalent ones
without updating, this is now incorrect as it changes the node id, and
thus invalidates the cache

At last, false_polymorphism used to modifiy the content of an analyse
during processing, this is invalid now that the analyses are cached