Rework dtype implementation
Do not use a dirty trick to represent dtype:

Former implementation returned a scalar of the according dtype, which
made it impossible to print the dtype.

Additionally, numpy_type was formerly used to harmonize between these
dtype and the one returned by numpy.uint8 and the likes.

There's now a proper type encapsulation that makes numpy_dtype useless,
and a type adapter types::dtype_t to get the proper dtype from a
scalar.