Drop big int support
Current implementation was partial, untested and buggy, to say the
least. And yet very few bug reports arise, which means the feature is
scarcely used.

With the forthcoming move to Python3, there will be no easy way to
distinguish system int and big int, so it's a dead end from Pythran
perspective any way.