dynamic tuple internal type
This type can be used as a tuple of elements of the same type, with a
size only known at runtime. It is an alternative to types::array,
introduced for situation when the tuple is created from an iterable with
unknown size.

This not a general solution of the list-to-tuple problem, but it helps.

This fixes #1152 and improves #1113.