Implement decent support for fixed-size list
a fixed-size list is just a std::array, the validity of the
list-to-static_list conversion is checked in
pythran/analyses/fixed_size_list.py and performed in
pythran/optimizations/list_to_tuple.py

There's now two flavour of types::array, namely array and static_list,
all based on a common template implementation of array_base. array is a
refinement of tuple while static_list is a refinement of list.

This fixes a previous issue in map/zip operators applied to fixed-sized
containers, where a tuple was generated instead of an list.

It also make it possible to build and update fixed-size containers out
of existing tuples without loosing the fixed-size information.