Handle the case where no strict aliases are found when checking default parameters
This situation is currently not well handled, so better be clear about it. It
turns out we could support situation where a call folds to a supported literal,
but that would imply calling constant folding a bit too early in the pipeline.

Or maybe we should call syntax checks later on?