Better document NDEBUG behavior
Use more asserts throughout the code to check for null pointers and buffer
overflow.

As this behavior is now documented, it should be somehow enforced by more
defensive coding under the protection of NDEBUG.

At some point, it may be interesting to separate checks that also exist in
Python (like IndexError or AssertionError) from internal checks...

This improves the situation wrt #416 but doesn't actually fixes it until the
check split isn't done.