Wrap new-delete in unique_ptr
new is used in few places to allocate temporary buffers, work and them
and delete them shortly after. Some of the operations done between the
new and delete can be not noexcept, so it is safer to use a unique_ptr
to ensure memory release even in case of error.