Improve typing wrt. combiner and lazy evaluation
Introduce a new type of node, LType, which delays its evaluation until the end
of the function typing, at which point we update its type and let the normal
evaluation process work.

This may introduce cyclic type dependencies but we solve them by providing an
initial, non-cyclic value for each of these types.

This may be a big step forward for better typing.