Correctly raise KeyError when getting an unset dict item
Previous version could not check whether it was running in load or store
more, and thus couldn't distinguish between the two modes.

Current version does this by forcing constness of dict at IR level when
they are not assigned to.

This should be done for all types, but that's too big of a change, so
pythonic::types::as_const currently only changes the constness of dict,
which is a hack, but hell...

Fix #2040