Import top-level submodules and functions lazily
This ensures that `import pythran` doesn't import a lot of
code which then starts to actually do things (like trying to import
numpy and scipy) which may fail or have undesirable side-effects
like emitting warnings.

This should also help with Cython, which does a:
```
try:
    import pythran
except ImportError:
    pass
```
which currently yields Pythran warnings even though the user doesn't
want to use Pythran but only has it installed in their environment.

In addition, it allows running `pip install -e . --no-build-isolation`
in SciPy twice in a row; something that was not working before because
of a recursive import issue (scipy needs pythran; pythran imports
scipy).

Closes gh-2126