Improve nested functions support
The transformation of a nested function into a partial_function is
only valid if all captured arguments are boxed, so that changes in the
parent are reflected in the partial_function.

We box each argument in a None-indexed dictionary, which is a
relatively lightweight approach that keeps Python semantic.
Unfortunately this stresses our type inference system a lot, so go the
hard way and improve it so that this works as expected.

Not that because they can be updated, boxed arguments are no longer
subject to false_polymorphism optimization.

Fix #2281